requirements.txt
ipython==8.12.3
matplotlib==3.8.0
mysql_connector_python==8.2.0
mysql_connector_repackaged==0.3.1
numpy==1.26.4
office365==0.3.15
Office365_REST_Python_Client==2.5.5
pandas==2.2.1
plotly==5.18.0
plotly_express==0.4.1
pygwalker==0.4.7
streamlit==1.31.1
streamlit_option_menu==0.3.12
